ci: Include dependencies in check workflow path filters
This allows us to detect breaking changes in dependent packages.

For example if we modify the Bank API Library then the checks for
the Capture SDK and Bank SDK will be executed, too. If these checks
fail then we know that our modifications in the Bank API Library
have introduced breaking changes.