feat(GiniCaptureSDK): Add `GiniCaptureNetworkService` protocol
This protocol allows us to extract the code related to network call
execution from the `DocumentService`.

By default we are using a default implementation which uses the
Gini Bank API Library.

Clients will be able to set their own implementation and thus easily
customise how network calls are executed without the need to deal
with the rest of the logic for managing document uploads.

PIA-2146