ci: Handle the case when a release tag already exists
Encountered the case where we had released newer versions from a different branch (beta releases of
version 3.0.0) but on 'main' the current version was older than the latest release tag.

By checking if the current version already has a release tag we can consider it released and
can skip creating another release tag for it.